# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Sentinel AML API
  version: 1.1.0
  description: Static API contract for the implementation in this repository. Basic
    authentication is stateless. Backend-configured users are admin, analyst and viewer;
    credentials come from environment variables and are never included here. All list
    responses mask customer identifiers and sensitive narrative fields. Numeric customer
    IDs support the authorized by-id detail endpoint. VIEWER detail views remain redacted.
    CSV and JSON batch submission returns 202 with a durable job to poll. Detection
    occurs atomically for each processed transaction. No external
    SAR filing API is implemented.
servers:
- url: http://localhost:8080/api/v1
  description: Local development; use HTTPS for deployment
security:
- basicAuth: []
paths:
...
  "/transactions":
    get:
      operationId: listTransactions
      summary: List transactions
      tags:
      - Transactions
      description: 'Allowed roles: ADMIN, ANALYST, VIEWER.'
      x-roles:
      - ADMIN
      - ANALYST
      - VIEWER
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PageTransactionResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
          default: 20
      - name: sort
        in: query
        schema:
          type: array
          items:
            type: string
        style: form
        explode: true
        description: Repeat for multiple sorts, e.g. sort=riskScore,desc.
    post:
      operationId: createTransaction
      summary: Create or ingest transaction
      tags:
      - Transactions
      description: 'Allowed roles: ADMIN.'
      x-roles:
      - ADMIN
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TransactionResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransactionRequest"
  "/transactions/{transactionRef}":
    get:
      operationId: getTransaction
      summary: Get transaction detail
      tags:
      - Transactions
      description: 'Allowed roles: ADMIN, ANALYST.'
      x-roles:
      - ADMIN
      - ANALYST
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TransactionResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: transactionRef
        in: path
        required: true
        schema:
          type: string
...
  "/transactions/batch":
    post:
      operationId: batchTransactions
      summary: Ingest a transaction batch
      tags:
      - Transactions
      description: 'ADMIN only. Returns a queued job; poll until COMPLETED or FAILED. Maximum 10000 records by default.'
      x-roles:
      - ADMIN
      responses:
        '202':
          description: Input persisted and queued; poll the Location URL for completion.
          headers:
            Location:
              description: Job status URL
              schema:
                type: string
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/IngestionJobResponse"
        '413':
          description: Input exceeds the configured size or record limit
        '429':
          description: Import queue is full; retry later
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 10000
              items:
                "$ref": "#/components/schemas/TransactionRequest"
  "/transactions/stream":
    post:
      operationId: streamTransaction
      summary: Ingest an idempotent transaction
      tags:
      - Transactions
      description: ADMIN only. First ingestion returns 201; replay of the same transactionRef
        returns its existing record with 200. Transaction persistence and detection
        commit atomically.
      x-roles:
      - ADMIN
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TransactionResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
        '200':
          description: Existing transaction replayed; no duplicate ingestion
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TransactionResponse"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransactionRequest"
...
  "/ingestion/customers/batch":
    post:
      operationId: importCustomersBatch
      summary: Import customers JSON batch
      tags:
      - Ingestion
      description: 'ADMIN only. Returns a queued job; poll until COMPLETED or FAILED. Maximum 10000 records by default.'
      x-roles:
      - ADMIN
      responses:
        '202':
          description: Input persisted and queued; poll the Location URL for completion.
          headers:
            Location:
              description: Job status URL
              schema:
                type: string
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/IngestionJobResponse"
        '413':
          description: Input exceeds the configured size or record limit
        '429':
          description: Import queue is full; retry later
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 10000
              items:
                "$ref": "#/components/schemas/CustomerRequest"
  "/ingestion/customers/csv":
    post:
      operationId: importCustomersCsv
      summary: Import customers CSV
      tags:
      - Ingestion
      description: ADMIN only. Asynchronous import with persisted progress and row-level errors. Multipart
        field name is file. CSV headers are snake_case; import customers before accounts
        before transactions.
      x-roles:
      - ADMIN
      responses:
        '202':
          description: Input persisted and queued; poll the Location URL for completion.
          headers:
            Location:
              description: Job status URL
              schema:
                type: string
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/IngestionJobResponse"
        '413':
          description: Input exceeds the configured size or record limit
        '429':
          description: Import queue is full; retry later
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
  "/ingestion/accounts/batch":
    post:
      operationId: importAccountsBatch
      summary: Import accounts JSON batch
      tags:
      - Ingestion
      description: 'ADMIN only. Returns a queued job; poll until COMPLETED or FAILED. Maximum 10000 records by default.'
      x-roles:
      - ADMIN
      responses:
        '202':
          description: Input persisted and queued; poll the Location URL for completion.
          headers:
            Location:
              description: Job status URL
              schema:
                type: string
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/IngestionJobResponse"
        '413':
          description: Input exceeds the configured size or record limit
        '429':
          description: Import queue is full; retry later
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 10000
              items:
                "$ref": "#/components/schemas/AccountRequest"
  "/ingestion/accounts/csv":
    post:
      operationId: importAccountsCsv
      summary: Import accounts CSV
      tags:
      - Ingestion
      description: ADMIN only. Asynchronous import with persisted progress and row-level errors. Multipart
        field name is file. CSV headers are snake_case; import customers before accounts
        before transactions.
      x-roles:
      - ADMIN
      responses:
        '202':
          description: Input persisted and queued; poll the Location URL for completion.
          headers:
            Location:
              description: Job status URL
              schema:
                type: string
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/IngestionJobResponse"
        '413':
          description: Input exceeds the configured size or record limit
        '429':
          description: Import queue is full; retry later
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
  "/ingestion/transactions/batch":
    post:
      operationId: importTransactionsBatch
      summary: Import transactions JSON batch
      tags:
      - Ingestion
      description: 'ADMIN only. Returns a queued job; poll until COMPLETED or FAILED. Maximum 10000 records by default.'
      x-roles:
      - ADMIN
      responses:
        '202':
          description: Input persisted and queued; poll the Location URL for completion.
          headers:
            Location:
              description: Job status URL
              schema:
                type: string
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/IngestionJobResponse"
        '413':
          description: Input exceeds the configured size or record limit
        '429':
          description: Import queue is full; retry later
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 10000
              items:
                "$ref": "#/components/schemas/TransactionRequest"
  "/ingestion/transactions/csv":
    post:
      operationId: importTransactionsCsv
      summary: Import transactions CSV
      tags:
      - Ingestion
      description: ADMIN only. Asynchronous import with persisted progress and row-level errors. Multipart
        field name is file. CSV headers are snake_case; import customers before accounts
        before transactions.
      x-roles:
      - ADMIN
      responses:
        '202':
          description: Input persisted and queued; poll the Location URL for completion.
          headers:
            Location:
              description: Job status URL
              schema:
                type: string
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/IngestionJobResponse"
        '413':
          description: Input exceeds the configured size or record limit
        '429':
          description: Import queue is full; retry later
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
  "/ingestion/errors":
    get:
      operationId: listIngestionErrors
      summary: List persisted ingestion errors
      tags:
      - Ingestion
      description: 'Allowed roles: ADMIN.'
      x-roles:
      - ADMIN
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PageIngestionErrorResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
          default: 20
      - name: sort
        in: query
        schema:
          type: array
          items:
            type: string
        style: form
        explode: true
        description: Repeat for multiple sorts, e.g. sort=riskScore,desc.
      - name: batchId
        in: query
        schema:
          type: string
...
  "/ingestion/jobs":
    get:
      operationId: listIngestionJobs
      summary: List recent ingestion jobs, newest first
      tags: [Ingestion]
      description: ADMIN only. List entries omit final result details; fetch an individual job for errors.
      x-roles: [ADMIN]
      parameters:
      - name: page
        in: query
        schema: {type: integer, minimum: 0, default: 0}
      - name: size
        in: query
        schema: {type: integer, minimum: 1, maximum: 500, default: 20}
      responses:
        '200':
          description: Persisted job history
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PageIngestionJobResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
  "/ingestion/jobs/{id}":
    get:
      operationId: getIngestionJob
      summary: Get import progress and final or partial results
      tags: [Ingestion]
      description: ADMIN only. COMPLETED may include rejected rows; FAILED may include committed rows. CSV totals are unknown until normal completion. Job ID also identifies the error-log batch.
      x-roles: [ADMIN]
      parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
      responses:
        '200':
          description: Current job state
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/IngestionJobResponse"
        '404':
          description: Unknown job ID
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    KycStatus:
      type: string
      enum:
      - PENDING
      - VERIFIED
      - REJECTED
      - EXPIRED
    IngestionSource:
      type: string
      enum:
      - API
      - BATCH
      - BULK_CSV
      - STREAM
...
    IngestionErrorType:
      type: string
      enum:
      - MALFORMED
      - VALIDATION
      - REFERENTIAL_INTEGRITY
      - DUPLICATE
      - PERSISTENCE
...
    AccountStatus:
      type: string
      enum:
      - ACTIVE
      - DORMANT
      - CLOSED
      - FROZEN
      - SUSPENDED
    RiskRating:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      - CRITICAL
...
    IngestionEntityType:
      type: string
      enum:
      - CUSTOMER
      - ACCOUNT
      - TRANSACTION
    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        transactionRef:
          type: string
        accountId:
          type: string
        customerId:
          type: string
        amount:
          type: number
        currency:
          type: string
        amountBase:
          type: number
        baseCurrency:
          type: string
        transactionType:
          type: string
        direction:
          type: string
        counterpartyName:
          type: string
        counterpartyAccount:
          type: string
        counterpartyCountry:
          type: string
        channel:
          type: string
        jurisdiction:
          type: string
        description:
          type: string
        transactionTime:
          type: string
          format: date-time
        ingestionSource:
          "$ref": "#/components/schemas/IngestionSource"
        createdAt:
          type: string
          format: date-time
    TransactionRequest:
      type: object
      properties:
        transactionRef:
          type: string
          maxLength: 100
          minLength: 1
        accountId:
          type: string
          maxLength: 64
          minLength: 1
        amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
          multipleOf: 0.01
          maximum: 100000000000000000
          exclusiveMaximum: true
          description: Positive decimal amount with at most 17 integer digits and
            2 fractional digits.
        currency:
          type: string
          minLength: 1
          pattern: "^[A-Za-z]{3}$"
        transactionType:
          type: string
          maxLength: 30
        direction:
          type: string
          maxLength: 10
        counterpartyName:
          type: string
          maxLength: 255
        counterpartyAccount:
          type: string
          maxLength: 100
        counterpartyCountry:
          type: string
          maxLength: 3
        channel:
          type: string
          maxLength: 50
        jurisdiction:
          type: string
          maxLength: 3
        description:
          type: string
          maxLength: 500
        transactionTime:
          type: string
          format: date-time
      required:
      - transactionRef
      - accountId
      - amount
      - currency
      - transactionTime
    IngestionJobResponse:
      type: object
      required: [jobId, batchId, entityType, status, processed, succeeded, failed, createdAt, statusUrl]
      properties:
        jobId: {type: string}
        batchId: {type: string, description: Same as jobId}
        entityType: {type: string, enum: [CUSTOMER, ACCOUNT, TRANSACTION]}
        sourceName: {type: string, nullable: true}
        status: {type: string, enum: [QUEUED, RUNNING, COMPLETED, FAILED]}
        totalRecords: {type: integer, nullable: true, description: Unknown for CSV until normal completion}
        processed: {type: integer, minimum: 0}
        succeeded: {type: integer, minimum: 0}
        failed: {type: integer, minimum: 0}
        createdAt: {type: string, format: date-time}
        startedAt: {type: string, format: date-time, nullable: true}
        finishedAt: {type: string, format: date-time, nullable: true}
        failureMessage: {type: string, nullable: true}
        statusUrl: {type: string}
        result:
          nullable: true
          description: Present on terminal job detail responses; partial on FAILED. Null while queued/running or in list entries.
          allOf:
          - "$ref": "#/components/schemas/IngestionResult"
    PageIngestionJobResponse:
      type: object
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/IngestionJobResponse"
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}
        last: {type: boolean}
    IngestionResult:
      type: object
      properties:
        batchId:
          type: string
        entityType:
          "$ref": "#/components/schemas/IngestionEntityType"
        sourceName:
          type: string
        totalRecords:
          type: integer
          format: int32
        succeeded:
          type: integer
          format: int32
        failed:
          type: integer
          format: int32
        errors:
          type: array
          items:
            "$ref": "#/components/schemas/IngestionErrorView"
      description: Synchronous per-record processing. totalRecords/succeeded/failed
        are full counts; errors contains at most 200 details. Successful records remain
        committed if other records fail.
...
    IngestionErrorView:
      type: object
      properties:
        sourceReference:
          type: string
        errorType:
          "$ref": "#/components/schemas/IngestionErrorType"
        message:
          type: string
        rawRecord:
          type: string
...
    AccountRequest:
      type: object
      properties:
        accountId:
          type: string
          maxLength: 64
          minLength: 1
        customerId:
          type: string
          maxLength: 64
          minLength: 1
        accountType:
          type: string
          maxLength: 30
        riskRating:
          "$ref": "#/components/schemas/RiskRating"
        accountStatus:
          "$ref": "#/components/schemas/AccountStatus"
        currency:
          type: string
          minLength: 1
          pattern: "^[A-Za-z]{3}$"
        openDate:
          type: string
          format: date
        closeDate:
          type: string
          format: date
        branchCode:
          type: string
          maxLength: 30
        branchCity:
          type: string
          maxLength: 100
        currentBalance:
          type: number
        avgMonthlyBalance6m:
          type: number
        creditLimit:
          type: number
        creditUtilizationPct:
          type: number
        overdraftEnabled:
          type: boolean
        cardType:
          type: string
          maxLength: 30
        jointAccount:
          type: boolean
        numLinkedDevices:
          type: integer
          format: int32
        mobileBankingEnrolled:
          type: boolean
        lastLoginDate:
          type: string
          format: date
        avgMonthlyTxnCount:
          type: integer
          format: int32
        accountTier:
          type: string
          maxLength: 30
      required:
      - accountId
      - customerId
      - currency
    CustomerRequest:
      type: object
      properties:
        customerId:
          type: string
          maxLength: 64
          minLength: 1
        firstName:
          type: string
          maxLength: 100
          minLength: 1
        lastName:
          type: string
          maxLength: 100
          minLength: 1
        gender:
          type: string
          maxLength: 20
        dateOfBirth:
          type: string
          format: date
        email:
          type: string
          maxLength: 255
          format: email
        phoneNumber:
          type: string
          maxLength: 50
        city:
          type: string
          maxLength: 100
        state:
          type: string
          maxLength: 100
        country:
          type: string
          maxLength: 3
        postalCode:
          type: string
          maxLength: 20
        occupation:
          type: string
          maxLength: 150
        annualIncome:
          type: number
        maritalStatus:
          type: string
          maxLength: 30
        educationLevel:
          type: string
          maxLength: 50
        employmentStatus:
          type: string
          maxLength: 30
        customerSince:
          type: string
          format: date
        customerSegment:
          type: string
          maxLength: 30
        kycStatus:
          "$ref": "#/components/schemas/KycStatus"
        riskRating:
          "$ref": "#/components/schemas/RiskRating"
        politicallyExposed:
          type: boolean
        preferredChannel:
          type: string
          maxLength: 50
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        numComplaintsLastYear:
          type: integer
          format: int32
      required:
      - customerId
      - firstName
      - lastName
...
    PageTransactionResponse:
      type: object
      required:
      - content
      - page
      - size
      - totalElements
      - totalPages
      - last
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/TransactionResponse"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
...
    PageIngestionErrorResponse:
      type: object
      required:
      - content
      - page
      - size
      - totalElements
      - totalPages
      - last
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/IngestionErrorResponse"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
...
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        errorType:
          type: string
        sourceReference:
          type: string
  responses:
    Unauthorized:
      description: Missing or invalid Basic credentials
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated role cannot access this resource
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Problem:
      description: Request failure; RFC 9457 problem details, including 400 validation,
        404 missing resource, 409 duplicate/conflict, 422 missing parent, or 500 unexpected
        failure.
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"